+++ src/components/iplSpace.vue
@@ -1,7 +1,7 @@
 <template>
     <div
         class="ipl-space"
-        :class="`color-${color}`"
+        :class="{[`color-${color}`]: true, clickable: clickable}"
     >
         <slot />
     </div>
@@ -15,11 +15,15 @@ export default defineComponent({
 
     props: {
         color: {
-            type: String as PropType<'dark' | 'light'>,
+            type: String as PropType<'dark' | 'light' | 'blue'>,
             default: 'dark',
             validator: (value: string): boolean => {
-                return ['dark', 'light'].includes(value);
+                return ['dark', 'light', 'blue'].includes(value);
             }
+        },
+        clickable: {
+            type: Boolean,
+            default: false
         }
     }
 });
@@ -39,6 +43,44 @@ export default defineComponent({
 
     &.color-dark {
         background-color: $background-primary;
+    }
+
+    &.color-blue {
+        background-color: $blue;
+    }
+
+    &.clickable {
+        user-select: none;
+        cursor: pointer;
+        transition-duration: $transition-duration-low;
+
+        &:hover {
+            &.color-light {
+                background-color: $background-secondary-hover;
+            }
+
+            &.color-dark {
+                background-color: $background-primary-hover;
+            }
+
+            &.color-blue {
+                background-color: $blue-hover;
+            }
+        }
+
+        &:active {
+            &.color-light {
+                background-color: $background-secondary-active;
+            }
+
+            &.color-dark {
+                background-color: $background-primary-active;
+            }
+
+            &.color-blue {
+                background-color: $blue-active;
+            }
+        }
     }
 }
 </style>