+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- Use inject/provide to pass validators to ipl-input instead of props
+- Export more helpers for validation, parsing colors and other small tasks
+- Add `clickable` prop to ipl-space
+
 # 1.2.1
 
 - Slightly change ipl-select padding to closer match ipl-input style